requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-limiter==2.4.0
prometheus-flask-exporter==0.18.4
opentelemetry-api==1.8.0
opentelemetry-sdk==1.8.0
opentelemetry-exporter-jaeger==1.8.0
gunicorn==20.1.0
python-dotenv==0.19.0
redis==3.5.3